pyproject.toml
[project]
name = "scene-agent-system"
version = "0.1.0"
description = "LangGraph agent pipeline driving a Three.js scene via AG-UI / CopilotKit"
requires-python = ">=3.11"
readme = "README.md"
authors = [{ name = "Eddie Nava" }]
license = { text = "MIT" }
dependencies = [
  "langgraph>=0.2",
  "langchain-anthropic>=0.2",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "ag-ui-protocol>=0.1",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "shared*"]
exclude = ["frontend*", "node_modules*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["agent/tests", "agent/evals"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "-m",
  "not e2e",
]
markers = [
  "e2e: end-to-end tests requiring API keys (deselected by default; run with `pytest -m e2e`)",
]

[tool.coverage.run]
source = ["agent"]
branch = true
omit = ["agent/tests/*", "agent/evals/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false
